requirements.txt
absl-py == 2.1.0
einops == 0.7.0
jax == 0.4.23
jaxlib == 0.4.23
numpy == 1.26.3
tensorboard == 2.17.0
tqdm == 4.66.1
